Moving test UI integration tests.
