+++ src/System.Windows.Forms/tests/UnitTests/System/Windows/Forms/Layout/AnchorLayoutTests.cs
@@ -72,43 +72,5 @@ namespace System.Windows.Forms.Layout.Tests
             Assert.NotNull(anchorInfo);
             Dispose(form, button);
         }
-
-        [WinFormsTheory]
-        [InlineData(AnchorStyles.Top, 120, 30, 20, 30)]
-        [InlineData(AnchorStyles.Left, 20, 180, 20, 30)]
-        [InlineData(AnchorStyles.Right, 220, 68, 20, 30)]
-        [InlineData(AnchorStyles.Bottom, 120, 330, 20, 30)]
-        [InlineData(AnchorStyles.Top | AnchorStyles.Left, 20, 30, 20, 30)]
-        [InlineData(AnchorStyles.Top | AnchorStyles.Right, 220, 30, 20, 30)]
-        [InlineData(AnchorStyles.Top | AnchorStyles.Bottom, 120, 30, 20, 330)]
-        [InlineData(AnchorStyles.Left | AnchorStyles.Right, 20, 180, 220, 30)]
-        [InlineData(AnchorStyles.Bottom | AnchorStyles.Right | AnchorStyles.Top | AnchorStyles.Left, 20, 30, 220, 330)]
-        public void ResizeAnchoredControlsParent_HanldeCreated_AnchorsApplied(AnchorStyles anchor, int x, int y, int width, int height)
-        {
-            var (form, button) = GetFormWithAnchoredButton();
-            button.Anchor = anchor;
-            DefaultLayout.AnchorInfo anchorInfo = DefaultLayout.GetAnchorInfo(button);
-            Assert.Null(anchorInfo);
-            form.Controls.Add(button);
-            form.Shown += OnFormShown;
-            form.ShowDialog();
-            Dispose(form, button);
-
-            void OnFormShown(object sender, EventArgs e)
-            {
-                Form formLocal = (Form)sender;
-                Control button1 = formLocal.Controls[0];
-
-                //Resize Form to compute button anchors.
-                formLocal.Size = new Size(400, 600);
-
-                Assert.Equal(x, button1.Bounds.X);
-                Assert.Equal(y, button1.Bounds.Y);
-                Assert.Equal(width, button1.Bounds.Width);
-                Assert.Equal(height, button1.Bounds.Height);
-
-                form.Close();
-            }
-        }
     }
 }
